fix: auto-advance handles milliseconds
Issue 15928

+++ AnkiDroid/src/main/java/com/ichi2/anki/reviewer/AutomaticAnswer.kt
@@ -252,12 +252,12 @@ class AutomaticAnswer(
 class AutomaticAnswerSettings(
     val answerAction: AutomaticAnswerAction = AutomaticAnswerAction.BURY_CARD,
     @get:JvmName("useTimer") val useTimer: Boolean = false,
-    private val secondsToShowQuestionFor: Int = 60,
-    private val secondsToShowAnswerFor: Int = 20
+    private val secondsToShowQuestionFor: Double = 60.0,
+    private val secondsToShowAnswerFor: Double = 20.0
 ) {
 
-    val millisecondsToShowQuestionFor = secondsToShowQuestionFor * 1000L
-    val millisecondsToShowAnswerFor = secondsToShowAnswerFor * 1000L
+    val millisecondsToShowQuestionFor = (secondsToShowQuestionFor * 1000L).toLong()
+    val millisecondsToShowAnswerFor = (secondsToShowAnswerFor * 1000L).toLong()
 
     // a wait of zero means auto-advance is disabled
     val autoAdvanceIfShowingAnswer; get() = secondsToShowAnswerFor > 0
